requirements.txt
colorama==0.4.6
numpy==2.2.3
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
paddleocr==2.10.0
supervision==0.25.0
ultralytics==8.3.88
playwright==1.41.2
python-dotenv